requirements.txt
anthropic>=0.55.0
flask>=3.0.0
python-dotenv>=1.0.0
